Update the Event class to expect a double for the 'num_attendee_rows' (which appears to be a recent API change).
- Also make sure that the connection gets released after making a service call.

+++ src/main/java/net/suaree/eventbrite/model/Event.java
@@ -61,7 +61,7 @@ public class Event extends EventData {
     private long capacity;
 
     @SerializedName("num_attendee_rows")
-    private long numAttendees;
+    private double numAttendees;
 
     private Organizer organizer;
 
@@ -194,7 +194,7 @@ public class Event extends EventData {
     }
 
     public long getNumAttendees() {
-        return numAttendees;
+        return (long)numAttendees;
     }
 
     public void setNumAttendees(long numAttendees) {
